moved logo into global folder
